+++ sismic/code/__init__.py
@@ -1,5 +1,6 @@
 from .evaluator import Evaluator
 from .dummy import DummyEvaluator
+from .enhanced import EnhancedEvaluator
 from .python import PythonEvaluator
 
-__all__ = ['Evaluator', 'DummyEvaluator', 'PythonEvaluator']
+__all__ = ['Evaluator', 'DummyEvaluator', 'EnhancedEvaluator', 'PythonEvaluator']
