+++ tests/test_code.py
@@ -1,7 +1,7 @@
 import pytest
 
 from sismic import code
-from sismic.code.python import FrozenContext
+from sismic.code.enhanced import FrozenContext
 from sismic.exceptions import CodeEvaluationError
 from sismic.interpreter import Event, InternalEvent, MetaEvent
 
@@ -93,7 +93,7 @@ class TestPythonEvaluator:
         events = evaluator._execute_code('send("hello", delay=5)')
         event = events[0]
         assert event == Event('hello', delay=5)
-        
+
     def test_notify(self, evaluator):
         events = evaluator._execute_code('notify("hello", x=1, y="world")')
         assert events == [MetaEvent('hello', x=1, y='world')]
